orangepi/meteo_subscriber/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
click==8.3.1
fastapi==0.127.1
h11==0.16.0
idna==3.11
jinja2==3.1.6
markupsafe==3.0.3
paho-mqtt==2.1.0
pydantic==2.12.5
pydantic-core==2.41.5
python-dotenv==1.2.1
starlette==0.50.0
typing-extensions==4.15.0
typing-inspection==0.4.2
uvicorn==0.40.0
